refactor(digitsd): extract conference callbacks into conference.go
Pure file-split refactor. No behavior change.

main.go shed 356 lines: every mesh (multi-party) phone.Callbacks
method that the controller drives via HandleHookFlash and the server's
ConferenceConnect (MutePeer, MigrateToMesh, TearDownPeer,
RequestConferenceMerge, AddMeshPeer, RemoveMeshPeer,
TearDownAllMeshPeers) plus the responder-side SDP setup helper
(setupMeshResponder) and the currentPeer lookup that picks between
the 2-party callPeer and the mesh's active peer for HOOK:FLASH
dispatch.

conference.go picks up the same code under package main with the same
receiver (*daemonCallbacks). Imports trimmed: webrtc/v4, sigclient,
owebrtc, phone, fmt, slog. main.go no longer imports the pion webrtc
package directly (all remaining webrtc.PeerConnectionState references
moved with the extracted code).

After this final extract, main.go is 2538 lines (down from the
~4700-line baseline before this refactor series). The team lead's
target was 'well under 2000, ideally closer to 1500.' We are above
that: the remaining mass is the main() function itself (the daemon
event loop), which would need restructuring rather than a cut+paste
move to split further. Calling it out for the close-out review.

Verification: go build, go test, golangci-lint, make build (Docker
arm64) all clean.
